requirements.txt
fastapi
uvicorn 
sqlalchemy 
pydantic 
sentence-transformers 
faiss-cpu 
psycopg2-binary